refactor(react-gui): give the APBS services a folder
`calcApbsPot.service.ts` was 618 lines covering everything one job does:
choosing two grids from the molecule's bounding box, writing the files two
external programs read, spawning and polling them, and loading the resulting
potential map into the scene.

`apbs/` separates the parts by what they are for -- `grid.ts` decides where
to solve, `inputFiles.ts` writes what the programs read, `naming.ts` names
the result, `jobs.ts` holds the in-flight jobs and the teardown every exit
path shares, and `run.ts` is the phase machine on top of them.

One behaviour fix falls out of moving the job counter into `jobs.ts`: ids
now come from `nextJobId()` rather than `++jobSeq` read across files, which
is the same guard `renderjob/` has -- a timestamped id lets two jobs started
in the same millisecond collide, and the loser's poll timer runs forever.
This counter was already monotonic, so nothing changes; it is now the only
way to get an id.

`worker/shared/apbsTypes.ts` stays where it is. The plan had it moving in,
but the settings pane, the dialog and the config context all read it -- it is
a boundary DTO, and `worker/server/` is not importable from the UI.

No behaviour change otherwise; the existing APBS tests pass with only their
import path moved.
